requirements.txt
# Core ML/DL Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1

# MLOps and Experiment Tracking
mlflow==2.7.1
protobuf==3.20.3

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# API and Web Frameworks
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
streamlit==1.26.0

# Data Processing and Storage
pyarrow==13.0.0
parquet==1.3.1
openpyxl==3.1.2

# Financial Data
yfinance==0.2.28
pandas-ta==0.3.14b0

# Statistical Tests
scipy==1.11.2

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0

# Development and Testing
pytest==7.4.2
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# Jupyter (optional, para exploración)
jupyter==1.0.0
notebook==7.0.3
ipywidgets==8.1.0

# Progress bars
tqdm==4.66.1

# Date/Time utilities
python-dateutil==2.8.2
